requirements.txt
# Django core
Django>=5.0,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
Pillow>=10.0,<11

# HTTP client (used by Meta Conversions API service)
requests>=2.31,<3

# Production server
gunicorn>=21.0,<23

# Database
dj-database-url>=2.1,<3
psycopg2-binary>=2.9,<3

# Static files
whitenoise>=6.6,<7

# Object storage (Cloudflare R2 via S3 compatibility)
django-storages>=1.14,<2
boto3>=1.34,<2
